requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-storage
google-cloud-secret-manager
vertexai
numpy
pypdf2
python-multipart
